refactor(harness-driver): split client.ts into spawn-request/broker-process
Decompose the 1199-line HarnessDriverClient file into two single-responsibility
siblings behind an unchanged public surface (client.ts now 931 lines):

- spawn-request.ts: spawn input -> broker /api/spawn request-body builders and
  the beforeAgentSpawn patch applier
- broker-process.ts: managed broker child-process lifecycle (API-URL wait,
  stdio drain, recent-output buffering, startup-error formatting, shutdown)

BrokerExitInfo moved into broker-process.ts and is re-exported from client.ts so
the package barrel surface is unchanged. Pure move: no behavior change. Verified
with build, full test suite (959 pass), and a harness-driver smoke test.
